Fixing bug in inputData handling (#2)
When previously attempting to debug performance prior to making the
initial commit to this repository, I was experimenting with increasing
the performance of the current code base.  As part of this, I
attempted to break apart inputData into its constituent elements prior
to passing it into other functions, assuming this would speed things
up.  Unfortunately, it did not and there were places where I forgot to
make changes that caused the code to break.

This change restores the code back to its original status, using
inputData as a struct, instead of using its constituent elements
separately.