Fill parameters, if missing
In newer versions of MATLAB, if you specify a (for example) gjr(1, 1)
model and while estimating it is determined that the coefficient is
zero, it will not place a value of zero into the result struct.
Instead, it will leave the placeholder empty.

This change inserts the zero back in where it belongs so that we have
proper starting values for the structural GARCH estimation.